TodoService/Project/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
